PartTask413: Fix bug, clean up and simplify script, add more unit tests

+++ dev_scripts/install/create_conda.py
@@ -1,20 +1,18 @@
 #!/usr/bin/env python
 
 """
-# Install the amp default environment:
-> create_conda.py --env_name develop --req_file dev_scripts/install/requirements/develop.txt --delete_env_if_exists
+# Install the `amp` default environment:
+> create_conda.py --env_name develop --req_file dev_scripts/install/requirements/develop.yaml --delete_env_if_exists
 
 # Install the `p1_develop` default environment:
-> create_conda.py --env_name p1_develop --req_file amp/dev_scripts/install/requirements/develop.txt --req_file dev_scripts/install/requirements/p1_develop.txt --delete_env_if_exists
+> create_conda.py --env_name p1_develop --req_file amp/dev_scripts/install/requirements/develop.yaml --req_file dev_scripts/install/requirements/p1_develop.yaml --delete_env_if_exists
 
 # Quick install to test the script:
 > create_conda.py --test_install -v DEBUG
 
-# Test the `develop` environment:
-> create_conda.py --env_name develop_test --req_file dev_scripts/install/requirements/develop.txt --delete_env_if_exists
-
-# Install pymc3 env:
-> create_conda.py --env_name pymc3 --req_file dev_scripts/install/requirements/pymc.txt -v DEBUG
+# Test the `develop` environment with a different name before switching the old
+# develop env:
+> create_conda.py --env_name develop_test --req_file dev_scripts/install/requirements/develop.yaml --delete_env_if_exists
 """
 
 import argparse
@@ -126,39 +124,34 @@ def _delete_conda_env(args, conda_env_name):
         "\n%s",
         prnt.frame("Delete old conda env '%s', if exists" % conda_env_name),
     )
-
-    if args.skip_delete_env:
-        _LOG.warning("Skipping")
-    else:
-        conda_env_dict, _ = hco.get_conda_info_envs()
-        conda_env_root = hco.get_conda_envs_dirs()[0]
-        conda_env_path = os.path.join(conda_env_root, conda_env_name)
-        if (
-            conda_env_name in conda_env_dict
-            or
-            # Sometimes conda is flaky and says that there is no env, even if
-            # the dir exists.
-            os.path.exists(conda_env_path)
-        ):
-            _LOG.warning("Conda env '%s' exists", conda_env_path)
-            if args.delete_env_if_exists:
-                # Back up the old environment.
-                # TODO(gp): Do this.
-                # Remove old dir to make conda happy.
-                _LOG.warning("Deleting conda env '%s'", conda_env_path)
-                # $CONDA remove -y -n $ENV_NAME --all
-                cmd = "conda deactivate; rm -rf %s" % conda_env_path
-                hco.conda_system(cmd, suppress_output=False)
-            else:
-                msg = (
-                    "Conda env '%s' already exists. You need to use"
-                    " --delete_env_if_exists to delete it" % conda_env_name
-                )
-                _LOG.error(msg)
-                sys.exit(-1)
+    conda_env_dict, _ = hco.get_conda_info_envs()
+    conda_env_root = hco.get_conda_envs_dirs()[0]
+    conda_env_path = os.path.join(conda_env_root, conda_env_name)
+    if (
+        conda_env_name in conda_env_dict
+        or
+        # Sometimes conda is flaky and says that there is no env, even if
+        # the dir exists.
+        os.path.exists(conda_env_path)
+    ):
+        _LOG.warning("Conda env '%s' exists", conda_env_path)
+        if args.delete_env_if_exists:
+            # Back up the old environment.
+            # TODO(gp): Do this.
+            # Remove old dir to make conda happy.
+            _LOG.warning("Deleting conda env '%s'", conda_env_path)
+            # $CONDA remove -y -n $ENV_NAME --all
+            cmd = "conda deactivate; rm -rf %s" % conda_env_path
+            hco.conda_system(cmd, suppress_output=False)
         else:
-            _LOG.warning("Skipping")
-    return conda_env_name
+            msg = (
+                "Conda env '%s' already exists. You need to use"
+                " --delete_env_if_exists to delete it" % conda_env_name
+            )
+            _LOG.error(msg)
+            sys.exit(-1)
+    else:
+        _LOG.warning("Skipping deleting environment")
 
 
 def _process_requirements_file(req_file):
@@ -178,12 +171,12 @@ def _process_requirements_file(req_file):
     dbg.dassert_exists(req_file)
     txt_tmp = io_.from_file(req_file, split=True)
     # Process.
-    for l in txt_tmp:
+    for line in txt_tmp:
         # TODO(gp): Can one do conditional builds for different machines?
         #  I don't think so.
-        if "# Not on Mac." in l:
+        if "# Not on Mac." in line:
             continue
-        txt.append(l)
+        txt.append(line)
     # Save file.
     txt = "\n".join(txt)
     dst_req_file = os.path.join(
@@ -203,37 +196,50 @@ def _process_requirements_files(req_files):
     return out_files
 
 
-def _install_conda_env(args, conda_env_name):
+def _create_conda_env(args, conda_env_name):
     """
-    Process requirements file and install conda.
+    Process requirements file and create conda env.
     """
     _LOG.info("\n%s", prnt.frame("Create new conda env '%s'" % conda_env_name))
-    if args.skip_install_env:
-        _LOG.warning("Skipping")
+    #
+    if args.test_install:
+        cmd = f"conda create --yes --name {conda_env_name} -c conda-forge"
     else:
         cmd = []
-        if args.yaml:
+        # Extract extensions.
+        extensions = set()
+        for req_file in args.req_file:
+            dbg.dassert_exists(req_file)
+            _, file_extension = os.path.splitext(req_file)
+            extensions.add(file_extension)
+        dbg.dassert_eq(
+            len(extensions),
+            1,
+            "There should be only one type of extension: found %s",
+            extensions,
+        )
+        extension = list(extensions)[0]
+        _LOG.debug("extension='%s'", extension)
+        dbg.dassert_in(extension, (".txt", ".yaml"), "Invalid req file extension")
+        if extension == ".yaml":
             cmd.append("conda env create")
         else:
+            dbg.dassert_eq(extension, ".txt")
             cmd.append("conda create")
             # Start installation without prompting the user.
             cmd.append("--yes")
             cmd.append("--name %s" % conda_env_name)
             # cmd.append("--override-channels")
-            # TODO(gp): Move to yaml?
             cmd.append("-c conda-forge")
-        if args.test_install:
-            pass
-        else:
-            req_files = args.req_file
-            tmp_req_files = _process_requirements_files(req_files)
-            # We leverage the fact that `conda create` can merge multiple
-            # requirements files.
-            cmd.append(" ".join(["--file %s" % f for f in tmp_req_files]))
+        req_files = args.req_file
+        tmp_req_files = _process_requirements_files(req_files)
+        # We leverage the fact that `conda create` can merge multiple
+        # requirements files.
+        cmd.append(" ".join(["--file %s" % f for f in tmp_req_files]))
         if args.python_version is not None:
             cmd.append("python=%s" % args.python_version)
         cmd = " ".join(cmd)
-        hco.conda_system(cmd, suppress_output=False)
+    hco.conda_system(cmd, suppress_output=False)
 
 
 def _test_conda_env(conda_env_name):
@@ -259,21 +265,23 @@ def _parse():
     )
     parser.add_argument("--delete_env_if_exists", action="store_true")
     parser.add_argument(
-        "--env_name", help="Environment name", default="develop", type=str
+        "--env_name", type=str, default=None, help="Environment name"
     )
-    parser.add_argument("--yaml", action="store_true")
     parser.add_argument(
         "--req_file", action="append", default=[], help="Requirements file"
     )
     # Debug options.
     parser.add_argument(
-        "--test_install", action="store_true", help="Just test the install step"
+        "--test_install",
+        action="store_true",
+        help="Test the install step without requirements",
     )
     parser.add_argument(
         "--python_version", default=None, type=str, action="store"
     )
     parser.add_argument("--skip_delete_env", action="store_true")
     parser.add_argument("--skip_install_env", action="store_true")
+    parser.add_argument("--skip_test_env", action="store_true")
     #
     parser.add_argument(
         "-v",
@@ -287,23 +295,37 @@ def _parse():
 
 def _main(parser):
     args = parser.parse_args()
-    dbg.dassert_is_not(args.env_name, None)
     dbg.init_logger(verb=args.log_level, use_exec_path=True)
+    #
     _LOG.info("\n%s", env.get_system_info(add_frame=True))
     dbg.dassert_exists(_REQUIREMENTS_DIR)
     dbg.dassert_exists(_CONDA_ENVS_DIR)
     #
     _set_conda_root_dir()
     #
-    conda_env_name = args.env_name
     if args.test_install:
         conda_env_name = "test_conda"
+    else:
+        dbg.dassert_is_not(args.env_name, None, "You need to specify an env_name")
+        conda_env_name = args.env_name
+        dbg.dassert_lte(
+            1, len(args.req_file), msg="You need to specify requirements files"
+        )
     #
-    _delete_conda_env(args, conda_env_name)
+    if args.skip_delete_env:
+        _LOG.warning("Skip delete conda env per user request")
+    else:
+        _delete_conda_env(args, conda_env_name)
     #
-    _install_conda_env(args, conda_env_name)
+    if args.skip_install_env:
+        _LOG.warning("Skip create conda env")
+    else:
+        _create_conda_env(args, conda_env_name)
     #
-    _test_conda_env(conda_env_name)
+    if args.skip_test_env:
+        _LOG.warning("Skip test conda env")
+    else:
+        _test_conda_env(conda_env_name)
     #
     _LOG.info("DONE")
